[Chore] #110 BorderView Class 활용하여 중복코드제거, BoardListTableViewCell 코드정리, 제네릭,프로토콜활용하여 같은 로직에 대한 중복코드 제거
