+++ GNUting/GNUting/Global/API/APIURLBase/APIURL.swift
@@ -98,8 +98,9 @@ enum EndPoint{
 }
 
 private extension URL{
-    static let baseURL = "http://203.255.15.32:14357/api/v1/"
-//    static let baseURL = "http://localhost:10001/api/v1/" // Test
+    static let baseURL = "http://203.255.15.32:1541/api/v1/" // Test
+//    static let baseURL = "http://203.255.15.32:14357/api/v1/"
+
     static func makeForEndpoint(endPoint : String) -> URL{
         URL(string: baseURL + endPoint)!
     }
@@ -109,6 +110,6 @@ private extension URL{
 }
 class BaseURL {
     static let shared = BaseURL()
-    let urlString = "http://203.255.15.32:14357/api/v1/"
-//    let urlString = "http://localhost:10001/api/v1/" // Test
+//    let urlString = "http://203.255.15.32:14357/api/v1/"
+    let urlString = "http://203.255.15.32:1541/api/v1/" // Test
 }
