+++ GNUting/GNUting/Presentation/MyPage/VC/UserWriteTextVC.swift
@@ -36,7 +36,7 @@ class UserWriteTextVC: BaseViewController {
     }()
     private lazy var dateBoardTableView : UITableView = {
        let tableView = UITableView()
-        tableView.register(DateBoardListTableViewCell.self, forCellReuseIdentifier: DateBoardListTableViewCell.identi)
+        tableView.register(BoardListTableViewCell.self, forCellReuseIdentifier: BoardListTableViewCell.identi)
         tableView.separatorStyle = .none
         tableView.showsVerticalScrollIndicator = false
         return tableView
@@ -101,8 +101,8 @@ extension UserWriteTextVC : UITableViewDataSource{
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: DateBoardListTableViewCell.identi, for: indexPath) as? DateBoardListTableViewCell else {return DateBoardListTableViewCell()}
-        cell.myPostSetCell(model: myPostList[indexPath.row])
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: BoardListTableViewCell.identi, for: indexPath) as? BoardListTableViewCell else {return BoardListTableViewCell()}
+        cell.setCell(model: myPostList[indexPath.row])
         
         return cell
     }
